+++ actix-files/CHANGES.md
@@ -8,6 +8,10 @@
 
 [#2615]: https://github.com/actix/actix-web/pull/2615
 [#3402]: https://github.com/actix/actix-web/issues/3402
+- Fix `NamedFile` panic when serving files with pre-UNIX epoch modification times. [#2748]
+
+[#2615]: https://github.com/actix/actix-web/pull/2615
+[#2748]: https://github.com/actix/actix-web/issues/2748
 
 ## 0.6.10